+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2022-08-22
+### Added
+- Dynamic URL Support
+### Changed
+
+- API based Product Page Pop UP modal
+- Fixed Checkout Session
+
+## [1.0.5] - 2022-07-06
+
+### Changed
+
+- Product Page Pop UP modal
+
+
+## [1.0.4] - 2022-06-22
+
+### Added
+
+- PHP 8.1 and Magento 2.4.4 Support
+
+### Changed
+
+- Setup Through Patch Data Interface
+
+
 ## [1.0.3] - 2021-11-09
 
 ### Added